move chunked structures
